String->symbol is banned.
This was a patch from John Cowan.  John explained: "String->symbol is
banned because in some implementations symbols are not garbage-
collected, so using it in the sandbox can consume resources to the
main program that it can't do anything about."

Thanks to CHIBA Masaomi for reporting this discrepancy between the
document and the sample implementation.